feature/ selectionLayer 로직을 새 파일로 분리함
