+++ src/components/PdfPage.vue
@@ -1,8 +1,16 @@
 <template>
     <div ref="$pdfPage" class="pdfPage card">
-        <canvas ref="$selectionLayer" class="selectionLayer"></canvas>
         <canvas ref="$pdfLayer" class="pdfLayer"></canvas>
-        <div ref="$textLayer" class="textLayer"></div>
+        <div
+            ref="$textLayer"
+            class="textLayer"
+            :data-page-index="pageIndex"
+        ></div>
+        <selection-layer
+            class="selectionLayer"
+            ref="$selectionLayer"
+            :pageIndex="pageIndex"
+        ></selection-layer>
     </div>
 </template>
 
@@ -15,8 +23,7 @@ import { usePdfStore } from '@/store/pdf';
 import { PageViewport, PDFPageProxy } from 'pdfjs-dist';
 import * as pdfjsLib from 'pdfjs-dist';
 import TOKEN from '@/constants/TOKEN';
-import { useSelectionStore } from '@/store/selection';
-import Line from '@/classes/Line';
+import SelectionLayer from './SelectionLayer.vue';
 
 const props = defineProps({
     pageIndex: {
@@ -25,85 +32,35 @@ const props = defineProps({
     },
 });
 const pdfStore = usePdfStore();
-const selectionStore = useSelectionStore();
 const $pdfPage = ref<HTMLDivElement>();
 const $pdfLayer = ref<HTMLCanvasElement>();
 const $textLayer = ref<HTMLDivElement>();
-const $selectionLayer = ref<HTMLCanvasElement>();
-const selectionLayerCtx = ref<CanvasRenderingContext2D | null>();
+const $selectionLayer = ref();
+const viewport = ref<PageViewport>();
+
 let page: PDFPageProxy;
-let viewport: PageViewport;
 let ctx: CanvasRenderingContext2D;
 
 onMounted(async () => {
     ctx = $pdfLayer.value?.getContext('2d') as CanvasRenderingContext2D;
     page = await pdfStore.getPage(props.pageIndex);
     const options = { scale: pdfStore.scale };
+    viewport.value = page.getViewport(options);
 
-    await renderPage(page, options);
+    await renderPage(page, viewport.value);
 
     if (!$textLayer.value) return;
-    if (!$selectionLayer.value) return;
-
-    selectionLayerCtx.value = $selectionLayer.value.getContext('2d');
     const nodes = [...$textLayer.value.childNodes];
 
     addTokenInfo(nodes);
 });
 
-selectionStore.$subscribe((_, state) => {
-    // document.addEventListener('selectionchange', () => {
-    // const selection = window.getSelection();
-    // const range = selection.getRangeAt(0);
-    const { range } = state;
-    if (!range || range.collapsed) {
-        clearSelectionLayer();
-        return;
-    }
-    const { startContainer, startOffset, endContainer, endOffset } = range;
-    if (
-        startContainer.nodeName !== '#text' ||
-        endContainer.nodeName !== '#text'
-    )
-        return;
-
-    clearSelectionLayer();
-
-    const selectedTokens = getSelectedTokens(range);
-    const lineMap = getLineMap(selectedTokens);
-    const $startToken = startContainer.parentElement;
-    const $endToken = endContainer.parentElement;
-
-    if (!$startToken || !$endToken) return;
-
-    const startLineNum = getLineNum($startToken);
-    const endLineNum = getLineNum($endToken);
-    const startLine = lineMap.get(startLineNum);
-    const endLine = lineMap.get(endLineNum);
-
-    if (!startLine || !endLine) return;
-
-    const startGap = getFirstLineStartOffset(startContainer, startOffset);
-    const endGap = getLastLineEndOffset(endContainer, endOffset);
-    startLine.setLeft(startLine.left + startGap);
-    endLine.setRight(endLine.right - endGap);
-
-    const selectedLines = [...lineMap.values()];
-    drawSelectionLines(selectedLines);
-});
 /**
  * page를 랜더링 합니다.
  * @param page pdf의 한 페이지에 해당하는 객체입니다.
  * @param options 랜더링 옵션 (scale 등)
  */
-async function renderPage(
-    page: PDFPageProxy,
-    options: {
-        scale: number;
-    }
-) {
-    viewport = await page.getViewport(options);
-
+async function renderPage(page: PDFPageProxy, viewport: PageViewport) {
     if (!page || !viewport || !ctx) {
         return;
     }
@@ -119,8 +76,8 @@ function setPageSize(viewport: PageViewport) {
     $pdfPage.value.style.height = viewport.height + 'px';
     $pdfLayer.value.width = viewport.width;
     $pdfLayer.value.height = viewport.height;
-    $selectionLayer.value.width = viewport.width;
-    $selectionLayer.value.height = viewport.height;
+    $selectionLayer.value.$el.width = viewport.width;
+    $selectionLayer.value.$el.height = viewport.height;
 }
 
 /**
@@ -163,91 +120,6 @@ async function renderTextLayer(page: PDFPageProxy, viewport: PageViewport) {
         container: $textLayer.value,
         viewport: viewport,
     });
-}
-function clearSelectionLayer() {
-    if (!selectionLayerCtx.value || !$selectionLayer.value) return;
-
-    const ctx = selectionLayerCtx.value;
-
-    ctx.clearRect(
-        0,
-        0,
-        $selectionLayer.value.width,
-        $selectionLayer.value.height
-    );
-}
-function drawSelectionLines(lines: Line[]) {
-    if (!selectionLayerCtx.value || !$selectionLayer.value) return;
-
-    const ctx = selectionLayerCtx.value;
-    lines.forEach((line) => {
-        const { left, top, right, bottom } = line;
-        ctx.beginPath();
-        ctx.rect(left, top, right - left, bottom - top);
-        ctx.fillStyle = 'red';
-        ctx.fill();
-    });
-}
-function getFirstLineStartOffset(startContainer: Node, startOffset: number) {
-    const beforeStartRange = new Range();
-    beforeStartRange.setStart(startContainer, 0);
-    beforeStartRange.setEnd(startContainer, startOffset);
-    const { width } = beforeStartRange.getClientRects()[0];
-
-    return width;
-}
-
-function getLastLineEndOffset(endContainer: Node, endOffset: number) {
-    const afterEndRange = new Range();
-    const wholeText = (endContainer as Text).wholeText;
-
-    afterEndRange.setStart(endContainer, endOffset);
-    afterEndRange.setEnd(endContainer, wholeText.length);
-
-    const { width } = afterEndRange.getClientRects()[0];
-
-    return width;
-}
-
-function getLineMap(tokens: HTMLElement[]) {
-    return tokens.reduce((map, $token) => {
-        const tokenInfo = $token.dataset;
-        const lineNum = parseInt(
-            tokenInfo[TOKEN.DATASET.LINE_NUM] as string,
-            10
-        );
-
-        if (!map.has(lineNum)) {
-            map.set(lineNum, new Line(lineNum));
-        }
-        map.get(lineNum)?.addToken($token);
-
-        return map;
-    }, new Map<number, Line>());
-}
-
-function getSelectedTokens(range: Range): HTMLElement[] {
-    console.log(range);
-    const selectedNodes = [...range.cloneContents().childNodes];
-    const commonAncestorContainer = range.commonAncestorContainer;
-
-    if (
-        commonAncestorContainer.nodeName === '#text' &&
-        commonAncestorContainer.parentElement
-    ) {
-        return [commonAncestorContainer.parentElement];
-    }
-
-    return selectedNodes
-        .filter((node) => node.nodeName === 'SPAN')
-        .map((node) => node as HTMLElement);
-}
-
-function getLineNum($token: HTMLElement): number {
-    const lineNumData = $token.dataset[TOKEN.DATASET.LINE_NUM];
-
-    if (!lineNumData) return -1;
-    return parseInt(lineNumData, 10);
 }
 
 function addTokenInfo(nodes: Node[]) {
